Consolidated session middleware
